+++ maslov.aleksey/T2/ioTypes.hpp
@@ -1,0 +1,12 @@
+#ifndef IO_TYPES_HPP
+#define IO_TYPES_HPP
+
+#include <iosfwd>
+#include "dataStruct.hpp"
+
+namespace maslov
+{
+  std::ostream & maslov::operator<<(std::ostream & out, const DataStruct & data);
+}
+
+#endif
