+++ maslov.aleksey/T2/main.cpp
@@ -1,50 +1,12 @@
-#include <string>
 #include <vector>
 #include <iostream>
 #include <iterator>
 #include <algorithm>
-#include <iomanip>
-#include "streamGuard.hpp"
+#include "dataStruct.hpp"
 
-struct DataStruct
-{
-  double key1;
-  char key2;
-  std::string key3;
-  bool operator<(const DataStruct & rhs) const;
-};
-bool DataStruct::operator<(const DataStruct & rhs) const
-{
-  if (key1 != rhs.key1)
-  {
-    return key1 < rhs.key1;
-  }
-  else if (key2 != rhs.key2)
-  {
-    return key2 < rhs.key2;
-  }
-  else
-  {
-    return key3.length() < rhs.key3.length();
-  }
-}
-std::ostream & operator<<(std::ostream & out, const DataStruct & data)
-{
-  std::ostream::sentry sentry(out);
-  if (!sentry)
-  {
-    return out;
-  }
-  maslov::StreamGuard guard(out);
-  out << "(";
-  out << ":key1 " << std::fixed << std::setprecision(6) << data.key1;
-  out << ":key2 '" << data.key2;
-  out << "':key3 \"" << data.key3;
-  out << "\":)";
-  return out;
-}
 int main()
 {
+  using namespace maslov;
   std::vector< DataStruct > data;
   while (!std::cin.eof())
   {
